data: service needs to be add grommunio group to service
The `grommunio` user has primary group `grommunio`.

uid=480(grommunio) gid=480(grommunio) groups=480(grommunio),486(systemd-journal),479(gromox)

By convention, the primary group is present in the supplementary
group list as well, but only by reference, which means that a call to
`setgid` changes that spot in the list.

The uwsgi process therefore lacks `grommunio` membership, but this is
needed to write to `/etc/gromox`. Explicitly add it as a secondary
group.